Updating docs-ui layout
