+++ docs-ui/src/app/(docs)/layout.module.css
@@ -2,5 +2,4 @@
   width: 100%;
   max-width: 960px;
   margin: 0 auto;
-  padding: 0 24px 64px;
 }
