+++ docs-ui/src/app/layout.module.css
@@ -1,0 +1,21 @@
+.container {
+  position: relative;
+  z-index: 20;
+  width: 100%;
+  margin-inline: 0 16px;
+  margin-top: 112px;
+  margin-bottom: 48px;
+  padding-inline: 40px;
+  padding-block: 48px;
+  background-color: white;
+  width: calc(100% - 32px);
+  border-radius: 0.5rem;
+}
+
+@media (min-width: 768px) {
+  .container {
+    width: calc(100% - 332px - 40px);
+    margin-left: 332px;
+    margin-right: 40px;
+  }
+}
