+++ docs-ui/src/components/Sidebar/index.tsx
@@ -1,57 +1,92 @@
+'use client';
+
 import styles from './Sidebar.module.css';
-import { Docs } from './docs';
-import { Playground } from './playground';
-import Link from 'next/link';
+import { components, overview, layoutComponents, theme } from '@/utils/data';
 import { ScrollArea } from '@base-ui-components/react/scroll-area';
+import Link from 'next/link';
+import { usePathname } from 'next/navigation';
+import { motion } from 'motion/react';
+import { Fragment } from 'react';
+
+const data = [
+  {
+    title: 'Overview',
+    content: overview,
+    url: '',
+  },
+  {
+    title: 'Theme',
+    content: theme,
+    url: '/theme',
+  },
+  {
+    title: 'Layout Components',
+    content: layoutComponents,
+    url: '/components',
+  },
+  {
+    title: 'Components',
+    content: components,
+    url: '/components',
+  },
+];
 
 export const Sidebar = () => {
+  const pathname = usePathname();
+  const isPlayground = pathname.includes('/playground');
+
   return (
     <div className={styles.sidebar}>
       <ScrollArea.Root className={styles.root}>
         <ScrollArea.Viewport className={styles.viewport}>
           <div className={styles.content}>
-            <div className={styles.logoContainer}>
-              <Link href="/">
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  width="133"
-                  height="24"
-                  viewBox="0 0 133 24"
-                  fill="none"
-                >
-                  <g clipPath="url(#clip0_13_214)">
-                    <path
-                      d="M13.5526 16.5056C12.4526 17.5563 10.959 18.1348 9.34564 18.1348C8.09243 18.1348 6.84752 17.7825 5.74444 17.1151L0 13.6401V15.0223L6.10088 18.714C7.12579 19.3333 8.25376 19.6332 9.34584 19.6332C10.7459 19.6332 12.0873 19.1403 13.0771 18.1947C13.8006 17.5034 14.2003 16.6747 14.1827 15.8704V15.7704C13.9976 16.0356 13.7865 16.2819 13.5526 16.5056V16.5056ZM13.5526 14.3224C12.4526 15.3729 10.959 15.9516 9.34564 15.9516C8.09243 15.9516 6.84752 15.5992 5.74444 14.9319L0 11.4569V12.839L6.10088 16.5304C7.12579 17.1499 8.25376 17.4508 9.34584 17.4508C10.7459 17.4508 12.0873 16.9579 13.0771 16.0122C13.8006 15.3209 14.2003 14.4923 14.1827 13.6882V13.5861C13.9978 13.8518 13.7867 14.0985 13.5526 14.3224V14.3224ZM13.5526 18.6899C12.4526 19.7404 10.959 20.3192 9.34564 20.3192C8.09243 20.3192 6.84752 19.9667 5.74444 19.2994L0 15.8235V17.2057L6.10088 20.8971C7.12579 21.5167 8.25376 21.8166 9.34584 21.8166C10.7459 21.8166 12.0873 21.3235 13.0771 20.3778C13.8006 19.6867 14.2003 18.8579 14.1827 18.054V17.9536C13.9976 18.219 13.7865 18.4653 13.5526 18.6891V18.6899ZM13.5526 20.8733C12.4526 21.9238 10.959 22.5025 9.34564 22.5025C8.09243 22.5025 6.84752 22.1501 5.74444 21.4828L0 18.0069V19.3891L6.10088 23.0805C7.12579 23.7 8.25376 24 9.34584 24C10.7459 24 12.0873 23.5071 13.0771 22.5612C13.8006 21.8701 14.2003 21.0412 14.1827 20.2374V20.1371C13.9976 20.4025 13.7865 20.6488 13.5526 20.8725V20.8733ZM18.3677 11.2799C18.3291 11.3184 18.2945 11.3588 18.2542 11.3971C17.8808 11.7468 17.46 12.0424 17.0041 12.2752C16.3408 12.6102 15.6151 12.8043 14.8728 12.8453H14.8699V14.3389H14.872C15.9209 14.272 16.9554 13.8722 17.7789 13.086C17.8238 13.0429 17.8634 12.9987 17.9061 12.9552C18.0678 12.7912 18.217 12.6154 18.3524 12.4292C18.4896 12.24 18.6079 12.0378 18.7054 11.8256C18.84 11.5352 18.9221 11.2234 18.948 10.9046C18.9532 10.8353 18.9595 10.766 18.9588 10.697V10.5742C18.9169 10.6342 18.8766 10.6947 18.8312 10.7543C18.6881 10.9392 18.5333 11.1148 18.3677 11.2799V11.2799ZM18.3696 9.09887C18.331 9.13735 18.2939 9.17583 18.2538 9.21431C17.8799 9.56293 17.4596 9.85854 17.0048 10.0928C16.5399 10.3265 16.0432 10.4909 15.5305 10.5809C15.2677 10.6292 15.0018 10.6585 14.7348 10.6685C14.7634 10.7589 14.7856 10.8507 14.8051 10.943C14.8519 11.164 14.8734 11.3895 14.8691 11.6153V12.1559H14.8722C15.3601 12.1262 15.8403 12.0214 16.296 11.8452C16.8469 11.63 17.3504 11.3098 17.7783 10.9024C17.8233 10.8593 17.8628 10.8153 17.9055 10.7716C18.0675 10.608 18.2169 10.4324 18.3524 10.2464C18.4897 10.0573 18.6079 9.85521 18.7054 9.64298C18.84 9.35266 18.9221 9.04093 18.948 8.72215C18.9532 8.65288 18.9595 8.58362 18.9588 8.51435V8.39891C18.9184 8.45663 18.876 8.51435 18.8322 8.57207C18.6894 8.75732 18.5349 8.93327 18.3696 9.09887V9.09887ZM18.3677 13.4633C18.3291 13.5018 18.2943 13.542 18.2542 13.5805C17.3451 14.4494 16.154 14.9569 14.8728 15.0287H14.8711V16.5223H14.8722C15.9193 16.4553 16.9549 16.0561 17.7783 15.2694C18.1998 14.8669 18.5072 14.4421 18.7046 14.0097C18.8392 13.7194 18.9213 13.4077 18.9472 13.0889C18.9524 13.0194 18.9588 12.9502 18.958 12.8811V12.7582C18.9161 12.8184 18.8758 12.8788 18.8304 12.9382C18.6876 13.123 18.533 13.2984 18.3677 13.4633ZM14.8728 17.2121V18.7057C15.9209 18.6387 16.9554 18.2397 17.7789 17.4528C18.5637 16.7024 18.9667 15.8783 18.9578 15.0645V14.9413C18.7519 15.2387 18.5161 15.5143 18.2542 15.7641C17.3449 16.6326 16.154 17.1403 14.8728 17.2121ZM16.9494 3.50652L11.1543 0L4.3986 6.45373L0 10.6556L6.10088 14.347C7.12579 14.9665 8.25376 15.2665 9.34584 15.2665C10.7459 15.2665 12.0873 14.7736 13.0771 13.8279C14.1536 12.7995 14.5108 11.4683 13.8465 10.3626C13.7387 10.183 13.6106 10.0163 13.4646 9.86578C13.8204 9.94486 14.1837 9.98505 14.5482 9.98565H14.5565C15.1517 9.98652 15.7419 9.87651 16.2966 9.66126C16.8475 9.44611 17.351 9.12587 17.7789 8.71849C17.8237 8.67597 17.8661 8.63287 17.9082 8.58978C18.0696 8.42604 18.2185 8.25046 18.3536 8.06452C18.4906 7.87557 18.6086 7.67359 18.7058 7.46153C18.8407 7.17228 18.9236 6.86163 18.9507 6.54377C19.034 5.45767 18.391 4.37772 16.9494 3.50652ZM11.3723 12.65C10.2876 13.6867 8.5415 13.8395 7.12097 12.9806L2.89953 10.4261L6.68831 6.80544L10.9806 9.40344C12.5281 10.3385 12.4428 11.628 11.3729 12.65H11.3723ZM16.0135 7.62353C14.9753 8.61575 13.2963 8.90416 11.8747 8.04509L7.90582 5.64295L11.5242 2.18337L15.4212 4.54203C16.8784 5.42284 17.0677 6.6167 16.0141 7.62353H16.0135Z"
-                      fill="url(#paint0_linear_13_214)"
-                    />
-                  </g>
-                  <path
-                    d="M30.128 5.22H34.61C35.606 5.22 36.428 5.358 37.076 5.634C37.724 5.91 38.198 6.294 38.498 6.786C38.81 7.266 38.966 7.824 38.966 8.46C38.966 8.988 38.852 9.468 38.624 9.9C38.396 10.32 38.072 10.668 37.652 10.944C37.232 11.208 36.746 11.394 36.194 11.502L36.176 11.286C36.86 11.37 37.442 11.556 37.922 11.844C38.414 12.12 38.786 12.474 39.038 12.906C39.29 13.338 39.416 13.83 39.416 14.382C39.416 15.066 39.254 15.684 38.93 16.236C38.618 16.776 38.114 17.208 37.418 17.532C36.734 17.844 35.864 18 34.808 18H30.128V5.22ZM34.664 15.948C35.144 15.948 35.54 15.882 35.852 15.75C36.176 15.618 36.416 15.426 36.572 15.174C36.728 14.91 36.806 14.586 36.806 14.202C36.806 13.83 36.722 13.518 36.554 13.266C36.398 13.002 36.158 12.804 35.834 12.672C35.522 12.54 35.138 12.474 34.682 12.474H32.666V15.948H34.664ZM34.394 10.44C34.85 10.44 35.228 10.374 35.528 10.242C35.84 10.11 36.068 9.924 36.212 9.684C36.356 9.432 36.428 9.144 36.428 8.82C36.428 8.484 36.35 8.202 36.194 7.974C36.05 7.746 35.822 7.572 35.51 7.452C35.21 7.332 34.826 7.272 34.358 7.272H32.666V10.44H34.394ZM43.2053 18.234C42.3653 18.234 41.6753 17.982 41.1353 17.478C40.5953 16.974 40.3253 16.32 40.3253 15.516C40.3253 14.604 40.6673 13.872 41.3513 13.32C42.0473 12.756 43.0733 12.474 44.4293 12.474H46.6793L45.9233 12.762V12.384C45.9233 11.916 45.7793 11.55 45.4913 11.286C45.2033 11.01 44.7293 10.872 44.0693 10.872C43.5053 10.872 42.9353 10.998 42.3593 11.25C41.7833 11.502 41.3033 11.814 40.9193 12.186V9.828C41.3033 9.54 41.8193 9.288 42.4673 9.072C43.1273 8.856 43.8053 8.748 44.5013 8.748C45.8093 8.748 46.7753 9.084 47.3993 9.756C48.0353 10.416 48.3533 11.322 48.3533 12.474V17.172C48.3533 17.316 48.3593 17.46 48.3713 17.604C48.3953 17.736 48.4253 17.868 48.4613 18H45.9953V15.3L46.0313 16.434C45.7793 16.974 45.4193 17.412 44.9513 17.748C44.4833 18.072 43.9013 18.234 43.2053 18.234ZM43.9073 16.47C44.3633 16.47 44.7653 16.302 45.1133 15.966C45.4733 15.618 45.7433 15.186 45.9233 14.67V14.022H44.6453C44.0213 14.022 43.5533 14.142 43.2413 14.382C42.9293 14.61 42.7733 14.934 42.7733 15.354C42.7733 15.69 42.8753 15.96 43.0793 16.164C43.2953 16.368 43.5713 16.47 43.9073 16.47ZM54.41 18.252C53.486 18.252 52.664 18.048 51.944 17.64C51.236 17.22 50.678 16.65 50.27 15.93C49.874 15.21 49.676 14.406 49.676 13.518C49.676 12.618 49.874 11.808 50.27 11.088C50.678 10.368 51.236 9.804 51.944 9.396C52.664 8.976 53.486 8.766 54.41 8.766C54.926 8.766 55.382 8.832 55.778 8.964C56.186 9.096 56.522 9.264 56.786 9.468V11.97C56.606 11.718 56.318 11.484 55.922 11.268C55.526 11.052 55.064 10.944 54.536 10.944C54.056 10.944 53.636 11.052 53.276 11.268C52.916 11.484 52.64 11.79 52.448 12.186C52.268 12.57 52.178 13.014 52.178 13.518C52.178 14.022 52.274 14.466 52.466 14.85C52.658 15.234 52.928 15.54 53.276 15.768C53.636 15.984 54.056 16.092 54.536 16.092C55.016 16.092 55.454 16.008 55.85 15.84C56.246 15.672 56.558 15.444 56.786 15.156V17.568C56.498 17.772 56.156 17.934 55.76 18.054C55.376 18.186 54.926 18.252 54.41 18.252ZM60.5674 13.086L64.2754 9.036H67.3534L63.2854 13.248L62.9974 13.482L60.5674 15.984V13.086ZM58.3354 5.22H60.7834V18H58.3354V5.22ZM62.2054 13.896L63.9514 12.33L67.5334 18H64.6534L62.2054 13.896ZM71.3489 18.27C70.6049 18.27 69.9269 18.162 69.3149 17.946C68.7029 17.73 68.2109 17.454 67.8389 17.118V14.634C68.2229 15.102 68.7269 15.486 69.3509 15.786C69.9749 16.074 70.6289 16.218 71.3129 16.218C71.7449 16.218 72.0749 16.152 72.3029 16.02C72.5309 15.888 72.6449 15.696 72.6449 15.444C72.6449 15.312 72.6089 15.192 72.5369 15.084C72.4649 14.976 72.3149 14.874 72.0869 14.778C71.8589 14.67 71.5229 14.562 71.0789 14.454L70.3409 14.274C69.5609 14.094 68.9369 13.794 68.4689 13.374C68.0009 12.942 67.7669 12.372 67.7669 11.664C67.7669 11.16 67.9109 10.686 68.1989 10.242C68.4989 9.786 68.9249 9.426 69.4769 9.162C70.0289 8.886 70.6649 8.748 71.3849 8.748C71.9969 8.748 72.5729 8.832 73.1129 9C73.6649 9.168 74.0849 9.36 74.3729 9.576V11.916C74.1449 11.64 73.7549 11.388 73.2029 11.16C72.6629 10.92 72.1049 10.8 71.5289 10.8C71.0729 10.8 70.7249 10.878 70.4849 11.034C70.2569 11.178 70.1429 11.352 70.1429 11.556C70.1429 11.7 70.1789 11.814 70.2509 11.898C70.3229 11.982 70.4729 12.066 70.7009 12.15C70.9289 12.234 71.2889 12.342 71.7809 12.474L72.5189 12.654C73.2869 12.846 73.8809 13.164 74.3009 13.608C74.7329 14.052 74.9489 14.628 74.9489 15.336C74.9489 15.84 74.8049 16.32 74.5169 16.776C74.2409 17.22 73.8329 17.58 73.2929 17.856C72.7529 18.132 72.1049 18.27 71.3489 18.27ZM79.9379 18.252C79.0499 18.252 78.3419 17.994 77.8139 17.478C77.2979 16.962 77.0399 16.2 77.0399 15.192V11.034H75.4199V9.846L79.5059 6.588H79.5239V9.036H81.6479V11.034H79.5239V14.832C79.5239 15.312 79.6259 15.654 79.8299 15.858C80.0339 16.062 80.3279 16.164 80.7119 16.164C80.9639 16.164 81.1919 16.122 81.3959 16.038C81.6119 15.954 81.8039 15.846 81.9719 15.714L81.5759 17.748C81.4319 17.892 81.2099 18.012 80.9099 18.108C80.6219 18.204 80.2979 18.252 79.9379 18.252ZM85.6565 18.234C84.8165 18.234 84.1265 17.982 83.5865 17.478C83.0465 16.974 82.7765 16.32 82.7765 15.516C82.7765 14.604 83.1185 13.872 83.8025 13.32C84.4985 12.756 85.5245 12.474 86.8805 12.474H89.1305L88.3745 12.762V12.384C88.3745 11.916 88.2305 11.55 87.9425 11.286C87.6545 11.01 87.1805 10.872 86.5205 10.872C85.9565 10.872 85.3865 10.998 84.8105 11.25C84.2345 11.502 83.7545 11.814 83.3705 12.186V9.828C83.7545 9.54 84.2705 9.288 84.9185 9.072C85.5785 8.856 86.2565 8.748 86.9525 8.748C88.2605 8.748 89.2265 9.084 89.8505 9.756C90.4865 10.416 90.8045 11.322 90.8045 12.474V17.172C90.8045 17.316 90.8105 17.46 90.8225 17.604C90.8465 17.736 90.8765 17.868 90.9125 18H88.4465V15.3L88.4825 16.434C88.2305 16.974 87.8705 17.412 87.4025 17.748C86.9345 18.072 86.3525 18.234 85.6565 18.234ZM86.3585 16.47C86.8145 16.47 87.2165 16.302 87.5645 15.966C87.9245 15.618 88.1945 15.186 88.3745 14.67V14.022H87.0965C86.4725 14.022 86.0045 14.142 85.6925 14.382C85.3805 14.61 85.2245 14.934 85.2245 15.354C85.2245 15.69 85.3265 15.96 85.5305 16.164C85.7465 16.368 86.0225 16.47 86.3585 16.47ZM97.0592 21.798C96.2312 21.798 95.4872 21.666 94.8272 21.402C94.1792 21.138 93.6572 20.778 93.2612 20.322C92.8772 19.866 92.6432 19.35 92.5592 18.774H95.1332C95.1932 19.098 95.3852 19.356 95.7092 19.548C96.0332 19.752 96.4592 19.854 96.9872 19.854C97.4192 19.854 97.7912 19.77 98.1032 19.602C98.4152 19.434 98.6552 19.164 98.8232 18.792C99.0032 18.432 99.0932 17.952 99.0932 17.352V14.436L99.4172 15.156C99.1172 15.972 98.6912 16.638 98.1392 17.154C97.5872 17.658 96.8672 17.91 95.9792 17.91C95.2472 17.91 94.5872 17.724 93.9992 17.352C93.4232 16.968 92.9732 16.434 92.6492 15.75C92.3252 15.054 92.1632 14.262 92.1632 13.374C92.1632 12.474 92.3252 11.676 92.6492 10.98C92.9732 10.284 93.4232 9.744 93.9992 9.36C94.5872 8.964 95.2472 8.766 95.9792 8.766C96.8792 8.766 97.6112 9.042 98.1752 9.594C98.7392 10.134 99.1832 10.842 99.5072 11.718L99.0932 12.24V9.036H101.559V17.082C101.559 18.09 101.379 18.948 101.019 19.656C100.659 20.364 100.143 20.898 99.4712 21.258C98.7992 21.618 97.9952 21.798 97.0592 21.798ZM96.6092 15.804C97.1972 15.804 97.7132 15.576 98.1572 15.12C98.6012 14.664 98.9372 14.076 99.1652 13.356C98.9372 12.612 98.6012 12.018 98.1572 11.574C97.7132 11.118 97.1972 10.89 96.6092 10.89C96.2252 10.89 95.8832 10.992 95.5832 11.196C95.2952 11.388 95.0672 11.67 94.8992 12.042C94.7432 12.414 94.6652 12.858 94.6652 13.374C94.6652 13.89 94.7432 14.328 94.8992 14.688C95.0672 15.048 95.2952 15.324 95.5832 15.516C95.8832 15.708 96.2252 15.804 96.6092 15.804ZM107.636 18.306C106.712 18.306 105.89 18.102 105.17 17.694C104.462 17.274 103.91 16.704 103.514 15.984C103.118 15.252 102.92 14.436 102.92 13.536C102.92 12.624 103.112 11.802 103.496 11.07C103.892 10.338 104.438 9.762 105.134 9.342C105.842 8.922 106.634 8.712 107.51 8.712C108.422 8.712 109.226 8.928 109.922 9.36C110.618 9.78 111.152 10.356 111.524 11.088C111.908 11.808 112.1 12.618 112.1 13.518C112.1 13.734 112.094 13.95 112.082 14.166H104.054V12.492H109.688L109.598 12.69C109.598 12.318 109.508 11.982 109.328 11.682C109.148 11.37 108.896 11.13 108.572 10.962C108.26 10.794 107.906 10.71 107.51 10.71C107.09 10.71 106.712 10.812 106.376 11.016C106.04 11.22 105.776 11.508 105.584 11.88C105.392 12.24 105.296 12.648 105.296 13.104V13.644C105.296 14.196 105.392 14.676 105.584 15.084C105.788 15.492 106.07 15.804 106.43 16.02C106.802 16.224 107.228 16.326 107.708 16.326C108.2 16.326 108.608 16.212 108.932 15.984C109.256 15.756 109.472 15.444 109.58 15.048H111.992C111.824 15.72 111.536 16.302 111.128 16.794C110.72 17.286 110.216 17.664 109.616 17.928C109.028 18.18 108.368 18.306 107.636 18.306ZM121.545 18.306C120.501 18.306 119.601 18.114 118.845 17.73C118.089 17.346 117.507 16.776 117.099 16.02C116.691 15.252 116.487 14.304 116.487 13.176V5.22H119.043V13.086C119.043 14.058 119.259 14.796 119.691 15.3C120.135 15.792 120.753 16.038 121.545 16.038C122.325 16.038 122.931 15.792 123.363 15.3C123.807 14.796 124.029 14.058 124.029 13.086V5.22H126.585V13.176C126.585 14.304 126.381 15.252 125.973 16.02C125.565 16.776 124.983 17.346 124.227 17.73C123.471 18.114 122.577 18.306 121.545 18.306ZM128.812 5.22H131.386V18H128.812V5.22Z"
-                    fill="black"
-                  />
-                  <defs>
-                    <linearGradient
-                      id="paint0_linear_13_214"
-                      x1="7.56907"
-                      y1="23.7803"
-                      x2="11.7317"
-                      y2="0.101547"
-                      gradientUnits="userSpaceOnUse"
-                    >
-                      <stop offset="0.07" stopColor="#36BAA2" />
-                      <stop offset="0.34" stopColor="#61DDC8" />
-                      <stop offset="0.53" stopColor="#7DF3E1" />
-                    </linearGradient>
-                    <clipPath id="clip0_13_214">
-                      <rect width="18.958" height="24" fill="white" />
-                    </clipPath>
-                  </defs>
-                </svg>
-              </Link>
-            </div>
             <div className={styles.menu}>
-              <Docs />
-              <Playground />
+              <motion.div
+                className={styles.section}
+                animate={{
+                  x: isPlayground ? -10 : 0,
+                  opacity: isPlayground ? 0 : 1,
+                  visibility: isPlayground ? 'hidden' : 'visible',
+                }}
+                initial={{
+                  x: isPlayground ? -10 : 0,
+                  opacity: isPlayground ? 0 : 1,
+                  visibility: isPlayground ? 'hidden' : 'visible',
+                }}
+                transition={{ duration: 0.2 }}
+              >
+                {data.map(section => {
+                  return (
+                    <Fragment key={section.title}>
+                      <div className={styles.sectionTitle}>{section.title}</div>
+
+                      {section.content.map(item => {
+                        const isActive =
+                          pathname === `${section.url}/${item.slug}`;
+
+                        return (
+                          <Link
+                            href={`${section.url}/${item.slug}`}
+                            key={item.slug}
+                            className={`${styles.line} ${
+                              isActive ? styles.active : ''
+                            }`}
+                          >
+                            <div className={styles.lineTitle}>{item.title}</div>
+                            <div className={styles.lineStatus}>
+                              {item.status === 'alpha' && 'Alpha'}
+                              {item.status === 'beta' && 'Beta'}
+                              {item.status === 'inProgress' && 'In Progress'}
+                              {item.status === 'stable' && 'Stable'}
+                              {item.status === 'deprecated' && 'Deprecated'}
+                            </div>
+                          </Link>
+                        );
+                      })}
+                    </Fragment>
+                  );
+                })}
+              </motion.div>
             </div>
           </div>
         </ScrollArea.Viewport>
