+++ docs-ui/src/components/Toolbar/Toolbar.tsx
@@ -1,0 +1,30 @@
+import { ThemeSelector } from './theme';
+import { ThemeNameSelector } from './theme-name';
+import styles from './Toolbar.module.css';
+import { Nav } from './nav';
+
+export const Toolbar = () => {
+  return (
+    <div className={styles.toolbar}>
+      <div className={styles.left}>
+        <svg
+          width="165"
+          height="35"
+          viewBox="0 0 165 35"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <path
+            d="M23.9672 14.1021C24.3495 13.8668 24.706 13.5918 25.0306 13.2816C25.0939 13.2207 25.1533 13.1598 25.2134 13.0988C25.4405 12.8677 25.65 12.6199 25.8401 12.3574C26.0335 12.0907 26.1997 11.8052 26.3362 11.5053C26.5264 11.0963 26.6431 10.6571 26.6811 10.2076C26.799 8.67342 25.8899 7.14791 23.8627 5.91753L15.7003 0.967529L6.18485 10.0842L-0.00976562 16.0185L8.5831 21.2328C9.95919 22.0753 11.5398 22.5246 13.1533 22.532C15.1251 22.532 17.0146 21.8357 18.4087 20.4993C19.9247 19.0466 20.428 17.1666 19.4927 15.6039C19.3403 15.3505 19.1599 15.115 18.9547 14.9021C19.4555 15.0137 19.967 15.0707 20.4802 15.0722H20.4936C21.3327 15.0736 22.1646 14.9183 22.9465 14.6141C23.3013 14.4742 23.6429 14.3028 23.9672 14.1021V14.1021ZM16.0081 18.8337C14.4794 20.2983 12.0202 20.5143 10.0184 19.3006L4.07383 15.6925L9.40995 10.5771L15.4566 14.2477C17.6357 15.5715 17.5154 17.3937 16.0081 18.8369V18.8337ZM16.7155 12.3289L11.1254 8.93848L16.2209 4.05337L21.7098 7.38528C23.7623 8.6299 24.0289 10.3168 22.5445 11.7371C21.7747 12.4654 20.7855 12.9187 19.7312 13.0263C18.6769 13.1339 17.6166 12.8898 16.7155 12.3321V12.3289ZM19.0781 27.3625C17.5297 28.8461 15.425 29.6634 13.1533 29.6634C11.3627 29.6562 9.60828 29.1584 8.08067 28.2242L-0.00976562 23.3209V25.2728L8.5831 30.4879C9.95919 31.3304 11.5398 31.7797 13.1533 31.7871C15.1251 31.7871 17.0146 31.09 18.4087 29.7544C19.4278 28.778 19.9912 27.6078 19.9659 26.4716V26.3299C19.7048 26.7046 19.4074 27.0525 19.0781 27.3688V27.3625ZM19.0781 24.2767C17.5297 25.761 15.425 26.5784 13.1533 26.5784C11.3627 26.5711 9.60828 26.0734 8.08067 25.1391L-0.00976562 20.2358V22.1886L8.5831 27.4037C9.95926 28.2459 11.5399 28.695 13.1533 28.7021C15.1251 28.7021 17.0146 28.0058 18.4087 26.6702C19.4278 25.693 19.9912 24.5235 19.9659 23.3865V23.2449C19.7052 23.6198 19.4078 23.9679 19.0781 24.2838V24.2767ZM19.0781 21.1908C17.5297 22.6752 15.425 23.4926 13.1533 23.4926C11.3627 23.4853 9.60828 22.9875 8.08067 22.0533L-0.00976562 17.1516V19.1043L8.5831 24.3194C9.95926 25.1616 11.5399 25.6107 13.1533 25.6178C15.1251 25.6178 17.0146 24.9215 18.4087 23.5851C19.4278 22.6087 19.9912 21.4393 19.9659 20.3023V20.1583C19.7052 20.5339 19.4078 20.8828 19.0781 21.1995V21.1908ZM25.7 23.2275C24.4086 24.4576 22.7188 25.1836 20.9375 25.2736V25.2736V27.395C22.4688 27.3098 23.9198 26.6824 25.0306 25.625C26.1383 24.5663 26.7041 23.4008 26.6922 22.2511V22.077C26.4016 22.496 26.0691 22.8844 25.7 23.2362V23.2275ZM19.0781 30.4444C17.5297 31.928 15.425 32.7461 13.1533 32.7461C11.3626 32.7389 9.60806 32.2409 8.08067 31.3061L-0.00976562 26.4051V28.3571L8.5831 33.5722C9.95919 34.4147 11.5398 34.864 13.1533 34.8714C15.1251 34.8714 17.0146 34.1743 18.4087 32.8387C19.4278 31.8623 19.9912 30.6921 19.9659 29.5558V29.4142C19.705 29.789 19.4076 30.137 19.0781 30.4531V30.4444ZM25.863 13.8117C25.8092 13.8663 25.7562 13.9209 25.7 13.9739C25.1735 14.4666 24.5816 14.8843 23.9411 15.2154C23.2865 15.5456 22.587 15.7778 21.8649 15.9046C21.4949 15.9731 21.1205 16.0147 20.7445 16.0288C20.784 16.1562 20.8157 16.2859 20.8434 16.4165C20.9094 16.7285 20.9394 17.0471 20.9328 17.366V18.1374H20.9367C21.6252 18.0962 22.3032 17.9486 22.9465 17.6999C23.7223 17.3964 24.4313 16.9443 25.0338 16.369C25.0979 16.3081 25.1533 16.2456 25.2134 16.1839C25.4405 15.9525 25.65 15.7044 25.8401 15.4417C26.0338 15.1746 26.2006 14.8889 26.3377 14.5887C26.5274 14.1786 26.6431 13.7383 26.6796 13.2879C26.6867 13.1906 26.6954 13.0925 26.6946 12.9952V12.8369C26.6376 12.916 26.5775 12.9999 26.5158 13.0814C26.3141 13.3414 26.096 13.5882 25.863 13.8204V13.8117ZM25.863 19.9771C25.8092 20.0325 25.7601 20.0886 25.7047 20.1432C24.4135 21.3735 22.7235 22.0993 20.9423 22.1886V22.1886V24.3091C22.4735 24.2236 23.9243 23.5962 25.0354 22.5391C25.5796 22.0346 26.0222 21.4305 26.3393 20.7596C26.5289 20.3495 26.6446 19.9091 26.6811 19.4588C26.6883 19.3607 26.697 19.2626 26.6962 19.1653V18.9912C26.6376 19.0766 26.5799 19.1621 26.5166 19.246C26.314 19.5066 26.0946 19.7537 25.8598 19.9858L25.863 19.9771ZM25.863 16.8912C25.8092 16.9466 25.7601 17.0028 25.7047 17.0574C25.442 17.3061 25.1611 17.5348 24.8645 17.7418C24.5715 17.9485 24.264 18.1336 23.9442 18.2957C23.0101 18.769 21.9879 19.0433 20.9423 19.1012H20.9383V21.2114C22.4695 21.1261 23.9204 20.4986 25.0314 19.4414C25.0955 19.3805 25.1509 19.318 25.211 19.2563C25.4382 19.0249 25.6476 18.7768 25.8377 18.5141C26.0313 18.2471 26.198 17.9617 26.3354 17.6619C26.525 17.2515 26.6407 16.8109 26.6772 16.3603C26.6843 16.263 26.693 16.1649 26.6922 16.0676V15.8935C26.6337 15.9789 26.5759 16.0644 26.5126 16.1483C26.3114 16.4131 26.0933 16.6647 25.8598 16.9015L25.863 16.8912ZM44.9319 9.86186C45.3507 9.86186 45.8015 9.86635 46.2841 9.87531C46.7668 9.88428 47.2447 9.91066 47.7178 9.95444C48.1631 9.99529 48.6054 10.0632 49.0424 10.1578C49.4134 10.2335 49.7692 10.371 50.0947 10.5645C50.7478 10.9604 51.3032 11.4983 51.7199 12.1383C52.1604 12.8145 52.3806 13.6416 52.3806 14.6196C52.3806 15.6503 52.1303 16.5412 51.6297 17.2924C51.1291 18.0436 50.4144 18.5998 49.4855 18.9611V19.0149C50.6871 19.2692 51.6068 19.812 52.2445 20.6433C52.8823 21.4746 53.2009 22.4874 53.2003 23.6817C53.2005 24.3866 53.071 25.0856 52.8182 25.7436C52.5664 26.4076 52.1804 27.0126 51.6843 27.5208C51.1545 28.0563 50.5232 28.4808 49.8273 28.7693C49.0888 29.0858 48.2374 29.2441 47.2732 29.2441H37.8519V9.87215L44.9319 9.86186ZM45.644 18.1912C47.121 18.1912 48.1815 17.9333 48.8256 17.4174C49.4696 16.9015 49.793 16.1285 49.7956 15.0983C49.7956 14.4125 49.6864 13.8698 49.4681 13.4699C49.2602 13.0837 48.9475 12.7639 48.5661 12.5473C48.1545 12.3247 47.703 12.1855 47.2376 12.1375C46.7076 12.0724 46.174 12.0407 45.64 12.0425H40.4495V18.2015L45.644 18.1912ZM46.9551 27.0634C48.1019 27.0634 48.9986 26.7559 49.6453 26.1408C50.292 25.5258 50.6127 24.6778 50.6075 23.597C50.6075 22.964 50.4893 22.4397 50.253 22.024C50.027 21.6199 49.7029 21.2792 49.3106 21.0334C48.8953 20.78 48.4369 20.6051 47.9584 20.5175C47.4369 20.4174 46.9069 20.3676 46.3759 20.3687H40.4463V27.0705L46.9551 27.0634ZM68.6976 29.1792C68.2962 29.4139 67.7407 29.5316 67.0312 29.5321C66.4299 29.5321 65.952 29.3649 65.5975 29.0305C65.243 28.696 65.0653 28.1487 65.0642 27.3886C64.4727 28.1149 63.707 28.6797 62.8384 29.0305C61.3812 29.5822 59.7915 29.6814 58.2769 29.3153C57.7411 29.188 57.2353 28.9574 56.7878 28.6364C56.3607 28.3248 56.0159 27.9139 55.783 27.4393C55.5377 26.9424 55.4148 26.341 55.4142 25.6352C55.4142 24.8393 55.5509 24.1881 55.8241 23.6817C56.0827 23.1917 56.4523 22.769 56.9033 22.4473C57.3682 22.1218 57.8856 21.8785 58.4328 21.7281C59.0067 21.5656 59.5941 21.4301 60.1949 21.3214C60.8321 21.1948 61.4387 21.0998 62.0148 21.0366C62.5263 20.9847 63.0331 20.894 63.5308 20.7652C63.9683 20.648 64.3133 20.4764 64.5689 20.2501C64.8245 20.0238 64.9511 19.6962 64.9511 19.2594C64.9511 18.753 64.8556 18.3461 64.6646 18.0385C64.485 17.7427 64.2306 17.4993 63.9272 17.3328C63.6118 17.1658 63.2695 17.0557 62.916 17.0076C62.5491 16.9543 62.1789 16.9279 61.8082 16.9284C60.8255 16.9284 60.0061 17.1131 59.3499 17.4823C58.6937 17.8515 58.3381 18.5523 58.2833 19.5846H55.9618C55.9982 18.7164 56.1802 17.9839 56.5077 17.3873C56.8222 16.8054 57.2718 16.3074 57.8188 15.9354C58.3878 15.5555 59.0223 15.2843 59.6901 15.1355C60.4295 14.969 61.1856 14.8878 61.9435 14.8934C62.579 14.8934 63.2136 14.9389 63.8425 15.0295C64.4404 15.1097 65.0189 15.2973 65.55 15.5833C66.0573 15.8596 66.4822 16.2656 66.7812 16.7599C67.0908 17.2647 67.2457 17.9241 67.2457 18.738V25.9383C67.2457 26.4795 67.2776 26.8764 67.3414 27.1291C67.4052 27.3818 67.6191 27.5081 67.9831 27.5081C68.2241 27.5026 68.4635 27.4662 68.6952 27.3997L68.6976 29.1792ZM64.9281 21.9892C64.6361 22.206 64.254 22.3627 63.7808 22.4639C63.3076 22.5652 62.8108 22.6436 62.2917 22.7069C61.7726 22.7702 61.2496 22.8422 60.7211 22.9229C60.2317 22.9923 59.7531 23.1236 59.2968 23.3137C58.8963 23.4802 58.5442 23.745 58.273 24.0836C58.0092 24.417 57.8774 24.8714 57.8774 25.4469C57.8665 25.7815 57.9469 26.1127 58.11 26.4051C58.2586 26.6597 58.4641 26.8763 58.7105 27.0381C58.9734 27.2073 59.2652 27.3265 59.5714 27.3894C59.9123 27.4627 60.2601 27.4988 60.6087 27.497C61.3736 27.497 62.0293 27.3936 62.5757 27.1869C63.067 27.012 63.5209 26.7463 63.9137 26.4035C64.2378 26.1178 64.4984 25.7675 64.6789 25.3749C64.8327 25.0494 64.9165 24.6952 64.9249 24.3352L64.9281 21.9892ZM79.2883 17.6627C78.7054 17.1748 77.9224 16.9305 76.9391 16.93C76.1009 16.93 75.3999 17.0838 74.836 17.3913C74.2916 17.6806 73.8271 18.1 73.4838 18.6122C73.1356 19.1455 72.8901 19.739 72.7598 20.3624C72.6129 21.035 72.5397 21.7217 72.5414 22.4101C72.5402 23.0503 72.6181 23.6881 72.7733 24.3091C72.9153 24.8948 73.1605 25.4504 73.4972 25.9502C73.8271 26.4317 74.2674 26.8273 74.7814 27.1038C75.3089 27.3934 75.9372 27.5379 76.6662 27.5374C77.8135 27.5374 78.7102 27.2391 79.3564 26.6425C80.0026 26.0459 80.3982 25.2048 80.5432 24.1192H82.917C82.6622 25.8552 82.0112 27.1937 80.9642 28.1348C79.9171 29.0758 78.4929 29.5461 76.6915 29.5455C75.617 29.5455 74.6656 29.3736 73.8375 29.0297C73.0425 28.7077 72.3333 28.2052 71.766 27.5619C71.1953 26.9016 70.7675 26.1303 70.5095 25.2966C70.2178 24.3625 70.0739 23.3886 70.083 22.4101C70.0757 21.414 70.2138 20.4222 70.4929 19.4659C70.7434 18.5951 71.1661 17.7834 71.7359 17.0788C72.2999 16.3916 73.014 15.8429 73.8232 15.4749C74.6625 15.0872 75.6383 14.8934 76.7508 14.8934C77.5113 14.8881 78.269 14.9839 79.0043 15.1782C79.6735 15.353 80.3037 15.6523 80.8621 16.0604C81.4039 16.4634 81.8551 16.9755 82.1866 17.5638C82.5321 18.1699 82.7508 18.8889 82.8426 19.7207H80.4388C80.2542 18.8345 79.8707 18.1485 79.2883 17.6627ZM87.9738 9.8761V21.3491L94.42 15.2178H97.5343L91.9347 20.3181L97.9442 29.2449H94.9937L90.1592 21.9188L87.9738 23.927V29.2449H85.6523V9.87215L87.9738 9.8761ZM101.468 26.1139C101.68 26.4612 101.976 26.7502 102.328 26.955C102.701 27.17 103.107 27.3214 103.53 27.4029C103.984 27.4943 104.446 27.5396 104.909 27.5382C105.293 27.5373 105.677 27.5109 106.057 27.459C106.438 27.4104 106.811 27.3097 107.165 27.16C107.492 27.0235 107.778 26.8056 107.998 26.527C108.216 26.2463 108.325 25.889 108.325 25.4548C108.325 24.8582 108.098 24.4062 107.642 24.0986C107.117 23.76 106.54 23.5081 105.935 23.3525C105.252 23.1626 104.51 22.9907 103.709 22.8366C102.947 22.6948 102.201 22.4767 101.483 22.1854C100.83 21.9245 100.247 21.5177 99.7759 20.9962C99.3202 20.4898 99.0925 19.7843 99.0931 18.8796C99.0931 18.1738 99.2513 17.5672 99.5678 17.0598C99.879 16.5606 100.299 16.1384 100.797 15.8246C101.319 15.4994 101.891 15.2613 102.49 15.1197C103.104 14.9692 103.733 14.8932 104.365 14.8934C105.109 14.8876 105.851 14.956 106.581 15.0975C107.216 15.2176 107.824 15.4568 108.37 15.8025C108.893 16.1414 109.321 16.6086 109.613 17.1595C109.922 17.7292 110.104 18.4389 110.159 19.2887H107.837C107.801 18.8366 107.683 18.4613 107.483 18.1628C107.286 17.8681 107.024 17.6223 106.717 17.4443C106.399 17.2593 106.054 17.1258 105.694 17.0487C105.321 16.9678 104.941 16.927 104.56 16.9268C104.208 16.9264 103.856 16.9528 103.508 17.006C103.173 17.0546 102.846 17.1505 102.538 17.2908C102.262 17.4137 102.018 17.5996 101.826 17.8336C101.644 18.0599 101.552 18.3537 101.552 18.7151C101.547 18.901 101.583 19.0857 101.656 19.2567C101.729 19.4277 101.838 19.5808 101.976 19.7057C102.292 19.9884 102.658 20.2093 103.055 20.3569C103.534 20.541 104.027 20.686 104.53 20.7905C105.077 20.9086 105.623 21.0307 106.169 21.1568C106.749 21.2834 107.318 21.4372 107.876 21.6181C108.4 21.7847 108.897 22.0275 109.35 22.3389C109.774 22.6336 110.129 23.018 110.388 23.4649C110.652 23.9175 110.784 24.4782 110.784 25.147C110.784 25.9979 110.606 26.7034 110.251 27.2636C109.901 27.8195 109.423 28.2843 108.858 28.6198C108.259 28.9729 107.602 29.2165 106.918 29.339C106.212 29.4733 105.494 29.5414 104.774 29.5424C104.021 29.5456 103.269 29.4636 102.534 29.2979C101.86 29.1518 101.22 28.8805 100.646 28.4979C100.097 28.1233 99.6467 27.6203 99.3352 27.0326C99.0076 26.4265 98.8256 25.6893 98.7892 24.8211H101.111C101.114 25.2746 101.238 25.7191 101.468 26.11V26.1139ZM119.386 15.2209V17.256H116.572V25.9612C116.566 26.1804 116.59 26.3993 116.641 26.6124C116.68 26.7658 116.772 26.9004 116.901 26.9922C117.061 27.0939 117.244 27.1545 117.433 27.1687C117.728 27.1989 118.025 27.2123 118.321 27.209H119.386V29.2441H117.61C117.093 29.2509 116.577 29.2099 116.068 29.1214C115.689 29.0625 115.332 28.9087 115.029 28.6744C114.748 28.4308 114.544 28.1101 114.442 27.7518C114.314 27.3541 114.25 26.8295 114.251 26.178V17.252H111.847V15.217H114.251V11.0115H116.572V15.217L119.386 15.2209ZM134.464 29.1895C134.062 29.4242 133.507 29.5419 132.798 29.5424C132.196 29.5424 131.718 29.3752 131.364 29.0407C131.009 28.7063 130.832 28.159 130.831 27.3989C130.239 28.1252 129.473 28.69 128.605 29.0407C127.148 29.5924 125.558 29.6917 124.043 29.3256C123.507 29.1983 123.002 28.9677 122.554 28.6467C122.129 28.3312 121.786 27.9168 121.557 27.4393C121.312 26.9424 121.189 26.341 121.188 25.6352C121.188 24.8393 121.325 24.1881 121.598 23.6817C121.857 23.1917 122.227 22.769 122.678 22.4473C123.142 22.1218 123.66 21.8785 124.207 21.7281C124.781 21.5656 125.368 21.4301 125.969 21.3214C126.606 21.1948 127.212 21.0998 127.789 21.0366C128.3 20.9847 128.807 20.894 129.305 20.7652C129.743 20.648 130.088 20.4764 130.343 20.2501C130.599 20.0238 130.725 19.6962 130.725 19.2594C130.725 18.753 130.63 18.3461 130.439 18.0385C130.259 17.7427 130.005 17.4993 129.701 17.3328C129.386 17.1658 129.044 17.0557 128.69 17.0076C128.323 16.9543 127.953 16.9279 127.582 16.9284C126.6 16.9284 125.78 17.1131 125.124 17.4823C124.468 17.8515 124.113 18.5523 124.059 19.5846H121.738C121.773 18.7164 121.955 17.9839 122.284 17.3873C122.597 16.8056 123.046 16.3077 123.592 15.9354C124.161 15.5555 124.796 15.2843 125.464 15.1355C126.203 14.9682 126.959 14.8862 127.717 14.891C128.352 14.8911 128.987 14.9366 129.616 15.0271C130.214 15.1073 130.792 15.2949 131.323 15.5809C131.83 15.858 132.254 16.2646 132.552 16.7591C132.862 17.2639 133.017 17.9233 133.017 18.7372V25.9375C133.017 26.4787 133.049 26.8756 133.112 27.1283C133.176 27.381 133.39 27.5073 133.754 27.5073C133.995 27.5018 134.235 27.4654 134.466 27.3989L134.464 29.1895ZM130.694 21.9995C130.402 22.2163 130.02 22.373 129.547 22.4742C129.074 22.5755 128.577 22.6538 128.058 22.7171C127.539 22.7804 127.016 22.8524 126.487 22.9332C125.999 23.0029 125.522 23.1342 125.067 23.324C124.667 23.4905 124.314 23.7553 124.043 24.0939C123.779 24.4273 123.648 24.8817 123.648 25.4572C123.637 25.7918 123.717 26.123 123.88 26.4154C124.029 26.67 124.234 26.8866 124.481 27.0484C124.744 27.2176 125.035 27.3367 125.342 27.3997C125.683 27.473 126.03 27.5091 126.379 27.5073C127.144 27.5073 127.8 27.4039 128.346 27.1971C128.837 27.0223 129.291 26.7566 129.684 26.4138C130.008 26.1281 130.269 25.7777 130.449 25.3852C130.603 25.0597 130.687 24.7055 130.695 24.3455L130.694 21.9995ZM147.219 33.205C146.164 34.3444 144.507 34.9141 142.248 34.9141C141.573 34.9122 140.9 34.8395 140.241 34.6973C139.587 34.564 138.96 34.3256 138.383 33.9915C137.835 33.6733 137.364 33.238 137.004 32.7168C136.639 32.1893 136.439 31.5471 136.402 30.7902H138.725C138.732 31.1779 138.871 31.5515 139.12 31.8488C139.371 32.1414 139.678 32.3811 140.022 32.5538C140.389 32.7393 140.78 32.8725 141.183 32.9495C141.579 33.0288 141.981 33.0699 142.385 33.0721C143.149 33.0721 143.796 32.9402 144.324 32.6765C144.841 32.4227 145.291 32.0498 145.635 31.5885C145.998 31.09 146.254 30.5218 146.386 29.9198C146.547 29.2168 146.625 28.4974 146.619 27.7763V26.8561H146.564C146.163 27.7244 145.557 28.3621 144.744 28.7693C143.943 29.1736 143.059 29.3826 142.162 29.3794C141.107 29.3794 140.187 29.1895 139.404 28.8097C138.645 28.4481 137.974 27.925 137.439 27.2771C136.899 26.612 136.495 25.8467 136.252 25.0252C135.984 24.141 135.851 23.2216 135.856 22.2978C135.861 21.4331 135.971 20.5721 136.184 19.7342C136.395 18.8759 136.767 18.0655 137.281 17.3462C137.812 16.6126 138.503 16.0095 139.302 15.5825C140.139 15.121 141.168 14.8902 142.388 14.8902C143.243 14.8813 144.087 15.0818 144.847 15.4741C145.594 15.8648 146.204 16.4754 146.594 17.2236H146.622V15.217H148.807V28.0501C148.805 30.3473 148.276 32.0656 147.219 33.205ZM144.255 26.8973C144.782 26.5913 145.227 26.163 145.553 25.6487C145.89 25.1101 146.135 24.5188 146.277 23.8993C146.43 23.2593 146.508 22.6037 146.509 21.9457C146.509 21.3325 146.436 20.7215 146.291 20.1258C146.154 19.5513 145.919 19.005 145.594 18.5117C145.277 18.0345 144.851 17.6394 144.351 17.3589C143.842 17.0693 143.223 16.9245 142.494 16.9245C141.747 16.9245 141.11 17.0648 140.582 17.3454C140.067 17.6137 139.622 17.9994 139.284 18.4713C138.938 18.9624 138.688 19.5147 138.547 20.0989C138.39 20.7292 138.313 21.3763 138.315 22.0256C138.315 22.6634 138.379 23.2996 138.506 23.9246C138.623 24.5324 138.845 25.1152 139.161 25.6471C139.464 26.1565 139.886 26.5849 140.391 26.8957C140.901 27.2122 141.538 27.3704 142.303 27.3704C143.068 27.3704 143.718 27.2127 144.255 26.8973V26.8973ZM162.296 28.3484C161.241 29.1444 159.911 29.5424 158.307 29.5424C157.178 29.5424 156.2 29.3615 155.371 28.9996C154.569 28.6571 153.854 28.1378 153.281 27.4804C152.702 26.7988 152.271 26.0042 152.015 25.147C151.725 24.187 151.565 23.1921 151.54 22.1894C151.529 21.1921 151.695 20.2007 152.032 19.2618C152.336 18.4099 152.803 17.625 153.406 16.9506C153.993 16.3028 154.71 15.7853 155.509 15.4314C156.348 15.0657 157.255 14.8824 158.169 14.8934C159.426 14.8934 160.469 15.1513 161.297 15.6672C162.104 16.1614 162.786 16.8344 163.291 17.6342C163.792 18.4321 164.144 19.3144 164.329 20.2382C164.515 21.1222 164.593 22.0256 164.561 22.9284H153.99C153.969 23.5197 154.043 24.1106 154.208 24.6786C154.365 25.2175 154.635 25.7168 155 26.1432C155.384 26.5819 155.861 26.9286 156.397 27.1576C156.961 27.4118 157.626 27.5384 158.391 27.5374C159.375 27.5374 160.181 27.3113 160.808 26.8593C161.436 26.4072 161.85 25.7199 162.051 24.7973H164.346C164.036 26.3708 163.353 27.5545 162.297 28.3484H162.296ZM161.736 19.3409C161.538 18.8681 161.25 18.4381 160.889 18.0749C160.302 17.5185 159.568 17.1432 158.773 16.994C157.978 16.8448 157.157 16.9281 156.409 17.2338C155.941 17.432 155.518 17.7229 155.166 18.0892C154.817 18.4554 154.54 18.883 154.347 19.3504C154.144 19.8381 154.024 20.356 153.992 20.883H162.104C162.07 20.3519 161.946 19.8305 161.737 19.3409H161.736ZM23.9672 14.1021C24.3495 13.8668 24.706 13.5918 25.0306 13.2816C25.0939 13.2207 25.1533 13.1598 25.2134 13.0988C25.4405 12.8677 25.65 12.6199 25.8401 12.3574C26.0335 12.0907 26.1997 11.8052 26.3362 11.5053C26.5264 11.0963 26.6431 10.6571 26.6811 10.2076C26.799 8.67342 25.8899 7.14791 23.8627 5.91753L15.7003 0.967529L6.18485 10.0842L-0.00976562 16.0185L8.5831 21.2328C9.95919 22.0753 11.5398 22.5246 13.1533 22.532C15.1251 22.532 17.0146 21.8357 18.4087 20.4993C19.9247 19.0466 20.428 17.1666 19.4927 15.6039C19.3403 15.3505 19.1599 15.115 18.9547 14.9021C19.4555 15.0137 19.967 15.0707 20.4802 15.0722H20.4936C21.3327 15.0736 22.1646 14.9183 22.9465 14.6141C23.3013 14.4742 23.6429 14.3028 23.9672 14.1021V14.1021ZM16.0081 18.8337C14.4794 20.2983 12.0202 20.5143 10.0184 19.3006L4.07383 15.6925L9.40995 10.5771L15.4566 14.2477C17.6357 15.5715 17.5154 17.3937 16.0081 18.8369V18.8337ZM16.7155 12.3289L11.1254 8.93848L16.2209 4.05337L21.7098 7.38528C23.7623 8.6299 24.0289 10.3168 22.5445 11.7371C21.7747 12.4654 20.7855 12.9187 19.7312 13.0263C18.6769 13.1339 17.6166 12.8898 16.7155 12.3321V12.3289ZM19.0781 27.3625C17.5297 28.8461 15.425 29.6634 13.1533 29.6634C11.3627 29.6562 9.60828 29.1584 8.08067 28.2242L-0.00976562 23.3209V25.2728L8.5831 30.4879C9.95919 31.3304 11.5398 31.7797 13.1533 31.7871C15.1251 31.7871 17.0146 31.09 18.4087 29.7544C19.4278 28.778 19.9912 27.6078 19.9659 26.4716V26.3299C19.7048 26.7046 19.4074 27.0525 19.0781 27.3688V27.3625ZM19.0781 24.2767C17.5297 25.761 15.425 26.5784 13.1533 26.5784C11.3627 26.5711 9.60828 26.0734 8.08067 25.1391L-0.00976562 20.2358V22.1886L8.5831 27.4037C9.95926 28.2459 11.5399 28.695 13.1533 28.7021C15.1251 28.7021 17.0146 28.0058 18.4087 26.6702C19.4278 25.693 19.9912 24.5235 19.9659 23.3865V23.2449C19.7052 23.6198 19.4078 23.9679 19.0781 24.2838V24.2767ZM19.0781 21.1908C17.5297 22.6752 15.425 23.4926 13.1533 23.4926C11.3627 23.4853 9.60828 22.9875 8.08067 22.0533L-0.00976562 17.1516V19.1043L8.5831 24.3194C9.95926 25.1616 11.5399 25.6107 13.1533 25.6178C15.1251 25.6178 17.0146 24.9215 18.4087 23.5851C19.4278 22.6087 19.9912 21.4393 19.9659 20.3023V20.1583C19.7052 20.5339 19.4078 20.8828 19.0781 21.1995V21.1908ZM25.7 23.2275C24.4086 24.4576 22.7188 25.1836 20.9375 25.2736V25.2736V27.395C22.4688 27.3098 23.9198 26.6824 25.0306 25.625C26.1383 24.5663 26.7041 23.4008 26.6922 22.2511V22.077C26.4016 22.496 26.0691 22.8844 25.7 23.2362V23.2275ZM19.0781 30.4444C17.5297 31.928 15.425 32.7461 13.1533 32.7461C11.3626 32.7389 9.60806 32.2409 8.08067 31.3061L-0.00976562 26.4051V28.3571L8.5831 33.5722C9.95919 34.4147 11.5398 34.864 13.1533 34.8714C15.1251 34.8714 17.0146 34.1743 18.4087 32.8387C19.4278 31.8623 19.9912 30.6921 19.9659 29.5558V29.4142C19.705 29.789 19.4076 30.137 19.0781 30.4531V30.4444ZM25.863 13.8117C25.8092 13.8663 25.7562 13.9209 25.7 13.9739C25.1735 14.4666 24.5816 14.8843 23.9411 15.2154C23.2865 15.5456 22.587 15.7778 21.8649 15.9046C21.4949 15.9731 21.1205 16.0147 20.7445 16.0288C20.784 16.1562 20.8157 16.2859 20.8434 16.4165C20.9094 16.7285 20.9394 17.0471 20.9328 17.366V18.1374H20.9367C21.6252 18.0962 22.3032 17.9486 22.9465 17.6999C23.7223 17.3964 24.4313 16.9443 25.0338 16.369C25.0979 16.3081 25.1533 16.2456 25.2134 16.1839C25.4405 15.9525 25.65 15.7044 25.8401 15.4417C26.0338 15.1746 26.2006 14.8889 26.3377 14.5887C26.5274 14.1786 26.6431 13.7383 26.6796 13.2879C26.6867 13.1906 26.6954 13.0925 26.6946 12.9952V12.8369C26.6376 12.916 26.5775 12.9999 26.5158 13.0814C26.3141 13.3414 26.096 13.5882 25.863 13.8204V13.8117ZM25.863 19.9771C25.8092 20.0325 25.7601 20.0886 25.7047 20.1432C24.4135 21.3735 22.7235 22.0993 20.9423 22.1886V22.1886V24.3091C22.4735 24.2236 23.9243 23.5962 25.0354 22.5391C25.5796 22.0346 26.0222 21.4305 26.3393 20.7596C26.5289 20.3495 26.6446 19.9091 26.6811 19.4588C26.6883 19.3607 26.697 19.2626 26.6962 19.1653V18.9912C26.6376 19.0766 26.5799 19.1621 26.5166 19.246C26.314 19.5066 26.0946 19.7537 25.8598 19.9858L25.863 19.9771ZM25.863 16.8912C25.8092 16.9466 25.7601 17.0028 25.7047 17.0574C25.442 17.3061 25.1611 17.5348 24.8645 17.7418C24.5715 17.9485 24.264 18.1336 23.9442 18.2957C23.0101 18.769 21.9879 19.0433 20.9423 19.1012H20.9383V21.2114C22.4695 21.1261 23.9204 20.4986 25.0314 19.4414C25.0955 19.3805 25.1509 19.318 25.211 19.2563C25.4382 19.0249 25.6476 18.7768 25.8377 18.5141C26.0313 18.2471 26.198 17.9617 26.3354 17.6619C26.525 17.2515 26.6407 16.8109 26.6772 16.3603C26.6843 16.263 26.693 16.1649 26.6922 16.0676V15.8935C26.6337 15.9789 26.5759 16.0644 26.5126 16.1483C26.3114 16.4131 26.0933 16.6647 25.8598 16.9015L25.863 16.8912Z"
+            fill="#121212"
+          />
+        </svg>
+        <div className={styles.name}>Button</div>
+      </div>
+      <div className={styles.actions}>
+        <ThemeNameSelector />
+        <ThemeSelector />
+      </div>
+    </div>
+  );
+};
