--- docs-ui/src/components/Toolbar/styles.module.css
@@ -1,22 +1,0 @@
-.toolbar {
-  position: sticky;
-  top: 16px;
-  left: 0;
-  right: 0;
-  border-radius: 0.5rem;
-  z-index: 10;
-  background-color: var(--panel);
-  height: 60px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0 1.5rem;
-  box-shadow: 0 0 20px 0 rgba(0, 0, 0, 0.025);
-  border: 1px solid var(--border);
-}
-
-.actions {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
